AI-Diet-Assistance-App-Backend/requirements.txt
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
celery==5.2.7
django-redis==5.2.0
pandas==1.4.4
django-celery-beat==2.4.0
psycopg2-binary==2.9.5